+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add `RedisClient.default_driver=`.
 - `#call` now takes an optional block to cast the return value.
 - Treat `#call` keyword arguments as Redis flags.
 - Fix `RedisClient#multi` returning some errors as values instead of raising them.